+++ Changelog.md
@@ -1,6 +1,7 @@
 Flan 1.7.6
 ================
 - Fix delete command using wrong permission node
+- Fabric: Support for EightsEconomyP
 
 Flan 1.7.5
 ================